+++ stable_baselines/sac/policies.py
@@ -4,17 +4,36 @@ from gym.spaces import Box
 
 from stable_baselines.common.policies import BasePolicy, nature_cnn, register_policy
 
-EPS = 1e-6
+EPS = 1e-6  # Avoid NaN (prevents division by zero or log of zero)
+# CAP the standard deviation of the actor
 LOG_STD_MAX = 2
 LOG_STD_MIN = -20
 
 
-def gaussian_likelihood(x, mu, log_std):
-    pre_sum = -0.5 * (((x - mu) / (tf.exp(log_std) + EPS)) ** 2 + 2 * log_std + np.log(2 * np.pi))
+def gaussian_logp(input_, mu, log_std):
+    """
+    Helper to computer log probability of a gaussian.
+    Here we assume this is a Diagonal Gaussian.
+
+    :param input_: (tf.Tensor)
+    :param mu: (tf.Tensor)
+    :param log_std: (tf.Tensor)
+    :return: (tf.Tensor)
+    """
+    pre_sum = -0.5 * (((input_ - mu) / (tf.exp(log_std) + EPS)) ** 2 + 2 * log_std + np.log(2 * np.pi))
     return tf.reduce_sum(pre_sum, axis=1)
 
 
 def mlp(input_ph, layers, activ_fn=tf.tanh, layer_norm=False):
+    """
+    Create a multi-layer fully connected neural network.
+
+    :param input_ph: (tf.placeholder)
+    :param layers: ([int]) Network architecture
+    :param activ_fn: (tf.function) Activation function
+    :param layer_norm: (bool) Whether to apply layer normalization or not
+    :return: (tf.Tensor)
+    """
     output = input_ph
     for i, layer_size in enumerate(layers):
         output = tf.layers.dense(output, layer_size, name='fc' + str(i))
@@ -31,13 +50,25 @@ def clip_but_pass_gradient(input_, lower=-1., upper=1.):
 
 
 def apply_squashing_func(mu, pi, logp_pi):
+    """
+    Squash the ouput of the gaussian distribution
+    and account for that in the log probability
+    The squashed mean is also returned for using
+    deterministic actions.
+
+    :param mu: (tf.Tensor) Mean of the gaussian
+    :param pi: (tf.Tensor) Output of the policy before squashing
+    :param logp_pi: (tf.Tensor) Log probability before squashing
+    :return: ([tf.Tensor])
+    """
+    # Squash the output
     mu = tf.tanh(mu)
     pi = tf.tanh(pi)
+    # OpenAI Variation:
     # To avoid evil machine precision error, strictly clip 1-pi**2 to [0,1] range.
     # logp_pi -= tf.reduce_sum(tf.log(clip_but_pass_gradient(1 - pi**2, l=0, u=1) + EPS), axis=1)
-    # Squash correction
+    # Squash correction (from original implementation)
     logp_pi -= tf.reduce_sum(tf.log(1 - pi ** 2 + EPS), axis=1)
-    # tf.reduce_sum(tf.log(1 - tf.tanh(actions) ** 2 + EPS), axis=1)
     return mu, pi, logp_pi
 
 
@@ -56,9 +87,10 @@ class SACPolicy(BasePolicy):
     :param scale: (bool) whether or not to scale the input
     """
 
-    def __init__(self, sess, ob_space, ac_space, n_env, n_steps, n_batch, n_lstm=256, reuse=False, scale=False):
-        super(SACPolicy, self).__init__(sess, ob_space, ac_space, n_env, n_steps, n_batch, n_lstm=n_lstm, reuse=reuse,
-                                        scale=scale)
+    def __init__(self, sess, ob_space, ac_space, n_env=1, n_steps=1, n_batch=None,
+                 n_lstm=256, reuse=False, scale=False):
+        super(SACPolicy, self).__init__(sess, ob_space, ac_space, n_env, n_steps, n_batch,
+                                        n_lstm=n_lstm, reuse=reuse, scale=scale)
         assert isinstance(ac_space, Box), "Error: the action space must be of type gym.spaces.Box"
         assert (np.abs(ac_space.low) == ac_space.high).all(), "Error: the action space low and high must be symmetric"
 
@@ -79,15 +111,18 @@ class SACPolicy(BasePolicy):
         """
         raise NotImplementedError
 
-    def make_critics(self, obs=None, action=None, reuse=False, scope="values_fn"):
+    def make_critics(self, obs=None, action=None, reuse=False,
+                     scope="values_fn", create_vf=True, create_qf=True):
         """
-        Creates the critics object
+        Creates the two Q-Values approximator along with the Value function
 
         :param obs: (TensorFlow Tensor) The observation placeholder (can be None for default placeholder)
-        :param action: (TensorFlow Tensor) The action placeholder (can be None for default placeholder)
+        :param action: (TensorFlow Tensor) The action placeholder
         :param reuse: (bool) whether or not to resue parameters
-        :param scope: (str) the scope name of the critic
-        :return: (TensorFlow Tensor) the output tensor
+        :param scope: (str) the scope name
+        :param create_vf: (bool) Whether to create Value fn or not
+        :param create_qf: (bool) Whether to create Q-Values fn or not
+        :return: ([tf.Tensor]) Mean, action and log probability
         """
         raise NotImplementedError
 
@@ -129,11 +164,13 @@ class FeedForwardPolicy(SACPolicy):
     :param cnn_extractor: (function (TensorFlow Tensor, ``**kwargs``): (TensorFlow Tensor)) the CNN feature extraction
     :param feature_extraction: (str) The feature extraction type ("cnn" or "mlp")
     :param layer_norm: (bool) enable layer normalisation
+    :param reg_weight: (float) Regularization loss weight for the policy parameters
     :param kwargs: (dict) Extra keyword arguments for the nature CNN feature extraction
     """
 
-    def __init__(self, sess, ob_space, ac_space, n_env, n_steps, n_batch, reuse=False, layers=None,
-                 cnn_extractor=nature_cnn, feature_extraction="cnn", layer_norm=False, **kwargs):
+    def __init__(self, sess, ob_space, ac_space, n_env=1, n_steps=1, n_batch=None, reuse=False, layers=None,
+                 cnn_extractor=nature_cnn, feature_extraction="cnn", reg_weight=0.0,
+                 layer_norm=False, **kwargs):
         super(FeedForwardPolicy, self).__init__(sess, ob_space, ac_space, n_env, n_steps, n_batch,
                                                 reuse=reuse, scale=(feature_extraction == "cnn"))
         self.layer_norm = layer_norm
@@ -144,6 +181,8 @@ class FeedForwardPolicy(SACPolicy):
         if layers is None:
             layers = [64, 64]
         self.layers = layers
+        self.reg_loss = None
+        self.reg_weight = reg_weight
 
         assert len(layers) >= 1, "Error: must have at least one hidden layer for the policy."
 
@@ -162,17 +201,26 @@ class FeedForwardPolicy(SACPolicy):
             pi_h = mlp(pi_h, self.layers, self.activ_fn, layer_norm=self.layer_norm)
 
             mu = tf.layers.dense(pi_h, self.ac_space.shape[0], activation=None)
+            # Important difference with SAC and other algo such as PPO:
+            # the std depends on the state, so we cannot use stable_baselines.common.distribution
             log_std = tf.layers.dense(pi_h, self.ac_space.shape[0], activation=None)
-        # OpenAI Variation
-        # activation = tf.tanh
+
+        # Regularize policy output (not used for now)
+        # reg_loss = self.reg_weight * 0.5 * tf.reduce_mean(log_std ** 2)
+        # reg_loss += self.reg_weight * 0.5 * tf.reduce_mean(mu ** 2)
+        # self.reg_loss = reg_loss
+
+        # OpenAI Variation to cap the standard deviation
+        # activation = tf.tanh # for log_std
         # log_std = LOG_STD_MIN + 0.5 * (LOG_STD_MAX - LOG_STD_MIN) * (log_std + 1)
         log_std = tf.clip_by_value(log_std, LOG_STD_MIN, LOG_STD_MAX)
 
         std = tf.exp(log_std)
         # Reparameterization trick
         pi = mu + tf.random_normal(tf.shape(mu)) * std
-        logp_pi = gaussian_likelihood(pi, mu, log_std)
+        logp_pi = gaussian_logp(pi, mu, log_std)
         # MISSING: reg params for log and mu
+        # Apply squashing and account for it in the probabilty
         mu, pi, logp_pi = apply_squashing_func(mu, pi, logp_pi)
         self.policy = pi
         self.deterministic_policy = mu
@@ -191,14 +239,17 @@ class FeedForwardPolicy(SACPolicy):
                 critics_h = tf.layers.flatten(obs)
 
             if create_vf:
+                # Value function
                 with tf.variable_scope('vf', reuse=reuse):
                     vf_h = mlp(critics_h, self.layers, self.activ_fn, layer_norm=self.layer_norm)
                     value_fn = tf.layers.dense(vf_h, 1, name="vf")
                 self.value_fn = value_fn
 
             if create_qf:
+                # Concatenate preprocessed state and action
                 qf_h = tf.concat([critics_h, action], axis=-1)
 
+                # Double Q values to reduce overestimation
                 with tf.variable_scope('qf1', reuse=reuse):
                     qf1_h = mlp(qf_h, self.layers, self.activ_fn, layer_norm=self.layer_norm)
                     qf1 = tf.layers.dense(qf1_h, 1, name="qf1")
@@ -235,7 +286,7 @@ class CnnPolicy(FeedForwardPolicy):
     :param _kwargs: (dict) Extra keyword arguments for the nature CNN feature extraction
     """
 
-    def __init__(self, sess, ob_space, ac_space, n_env, n_steps, n_batch, reuse=False, **_kwargs):
+    def __init__(self, sess, ob_space, ac_space, n_env=1, n_steps=1, n_batch=None, reuse=False, **_kwargs):
         super(CnnPolicy, self).__init__(sess, ob_space, ac_space, n_env, n_steps, n_batch, reuse,
                                         feature_extraction="cnn", **_kwargs)
 
@@ -254,7 +305,7 @@ class LnCnnPolicy(FeedForwardPolicy):
     :param _kwargs: (dict) Extra keyword arguments for the nature CNN feature extraction
     """
 
-    def __init__(self, sess, ob_space, ac_space, n_env, n_steps, n_batch, reuse=False, **_kwargs):
+    def __init__(self, sess, ob_space, ac_space, n_env=1, n_steps=1, n_batch=None, reuse=False, **_kwargs):
         super(LnCnnPolicy, self).__init__(sess, ob_space, ac_space, n_env, n_steps, n_batch, reuse,
                                           feature_extraction="cnn", layer_norm=True, **_kwargs)
 
@@ -273,7 +324,7 @@ class MlpPolicy(FeedForwardPolicy):
     :param _kwargs: (dict) Extra keyword arguments for the nature CNN feature extraction
     """
 
-    def __init__(self, sess, ob_space, ac_space, n_env, n_steps, n_batch, reuse=False, **_kwargs):
+    def __init__(self, sess, ob_space, ac_space, n_env=1, n_steps=1, n_batch=None, reuse=False, **_kwargs):
         super(MlpPolicy, self).__init__(sess, ob_space, ac_space, n_env, n_steps, n_batch, reuse,
                                         feature_extraction="mlp", **_kwargs)
 
@@ -292,7 +343,7 @@ class LnMlpPolicy(FeedForwardPolicy):
     :param _kwargs: (dict) Extra keyword arguments for the nature CNN feature extraction
     """
 
-    def __init__(self, sess, ob_space, ac_space, n_env, n_steps, n_batch, reuse=False, **_kwargs):
+    def __init__(self, sess, ob_space, ac_space, n_env=1, n_steps=1, n_batch=None, reuse=False, **_kwargs):
         super(LnMlpPolicy, self).__init__(sess, ob_space, ac_space, n_env, n_steps, n_batch, reuse,
                                           feature_extraction="mlp", layer_norm=True, **_kwargs)